SwitchCheckModal purification
